content/media/video/test/test_can_play_type.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=469247
-->
<head>
  <title>Test for Bug 469247</title>
  <script type="application/javascript" src="/MochiKit/MochiKit.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=469247">Mozill
a Bug 469247</a>
<p id="display"></p>
<div id="content" style="display: none">
</div>

<video id="v" onerror="event.stopPropagation();"></video>

<pre id="test">
<script type="application/javascript">

var v = document.getElementById('v');

function check(type, expected) {
  is(v.canPlayType(type), expected, type);
}

// Invalid types
check("foo/bar", "no");
check("", "no");
check("!!!", "no");

</script>
</pre>
</body>
</html>
